+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT
 - Add type `above` in tiled to render above components
 - update `tiledjsonreader`
+- improvements in `Camera`
 
 # 1.0.2
 - Downgrade flame version to 1.0.0-rc9