+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add type `above` in tiled to render above components
 - update `tiledjsonreader`
 - improvements in `Camera`
+- others improvements
 
 # 1.0.2
 - Downgrade flame version to 1.0.0-rc9